+++ src/StorageContext.tsx
@@ -1,13 +1,7 @@
 import { createContext } from 'react';
-
-export interface StorageConfig {
-  storage: Storage;
-  serializer: (data: any) => string;
-  parser: (raw: string) => any;
-}
+import { StorageConfig } from './StorageConfig';
 
 export const storageContextDefaultValue = {
-  storage: localStorage,
   serializer: JSON.stringify,
   parser: JSON.parse,
 };