+++ src/index.ts
@@ -1,3 +1,4 @@
+export * from './StorageConfig';
 export * from './StorageProvider';
 export * from './useLocalStorage';
 export * from './useSessionStorage';